gaijiekou

+++ app/src/main/java/com/example/sourcewords/ui/mine/view/MineFragment.java
@@ -116,7 +116,7 @@ public class MineFragment extends Fragment {
             public void success(PlanItem myplan) {
                 if(myplan.getData().getPlans() != null){
                     mine_progress.setText(myplan.getData().getPlans().get(0).getPercent() + "%");
-                    mine_bar.setProgress(myplan.getData().getPlans().get(0).getPercent());
+                    mine_bar.setProgress(myplan.getData().getPlans().get(0).getPercent().intValue());
                 }else{
                     mine_progress.setText("0%");
                     mine_bar.setProgress(0);
